ref: move rejection handshake tests
